requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.30.0
anthropic>=0.28.0
pypdf2>=3.0.1
pdfplumber>=0.10.3
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.8.0
tiktoken>=0.7.0
sentence-transformers>=3.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
gunicorn>=22.0.0
werkzeug>=3.0.0
